pyproject.toml
[project]
name = "martyrologe"
version = "0.1.0"
description = "Données structurées pour un martyrologe."
author = "Timothé Watteau"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.50.1",
    "numpy>=2.3.5",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
]
